+++ mc_recipe_img/cli.py
@@ -1,8 +1,7 @@
 import shutil
 from enum import Enum
 from pathlib import Path
-from typing import Annotated, Never, cast
-from zipfile import ZipFile
+from typing import Annotated, Never
 
 import typer
 from PIL import Image
@@ -13,7 +12,7 @@ from rich.theme import Theme
 from tabulate2 import tabulate
 
 from mc_recipe_img import FILE_CACHE, USER_CACHE_DIR, init_logger, logger, mc
-from mc_recipe_img.util import group_as_dict, parse_envvar_paths, partitioned
+from mc_recipe_img.util import parse_envvar_paths, partitioned
 
 
 class LogLevel(str, Enum):
@@ -149,30 +148,7 @@ def run(
     #region ANALYZE RECIPES, COLLECT TEXTURES
 
     pack: mc.Datapack = mc.Datapack(datapack_dir, mc_versions)
-    texture_map: dict[str, Path] = pack.find_required_textures(*textures_sources)
-
-    # Extract textures from JAR where necessary
-    for jar_path, namelist in group_as_dict(
-            (t for t in texture_map.values() if '.jar::assets' in str(t)),
-            lambda t: cast(tuple[str, str], tuple(str(t).split('::'))),
-        ).items():
-        ext_dir: Path = USER_CACHE_DIR / f'jar_extracted/{Path(jar_path).stem}'
-        if not ext_dir.exists():
-            ext_dir.mkdir(parents=True)
-        with ZipFile(jar_path) as jar:
-            logger.info(f'Extracting {len(namelist)} textures from {jar_path} to: {ext_dir}')
-            new: int = 0
-            for name in namelist:
-                dest: Path = ext_dir / name
-                texture_map[f'minecraft:{Path(name).stem}'] = dest
-                if dest.is_file():
-                    logger.debug(f'Already extracted: {dest}')
-                    continue
-                logger.debug(f'Extracting: {name} -> {dest}')
-                new += 1
-                jar.extract(name, ext_dir)
-            logger.info(f'{new} new textures, {len(namelist) - new} already extracted')
-            del name
+    texture_map: dict[str, Path] = pack.get_texture_map(*textures_sources, extract=True)
 
     #endregion ANALYZE RECIPES, COLLECT TEXTURES
 
@@ -180,7 +156,7 @@ def run(
 
     logger.debug(f'Texture map:\n{tabulate(texture_map.items())}')
 
-    rendered: dict[str, Image.Image] = pack.render_recipes(texture_map)
+    rendered: dict[str, Image.Image] = pack.render_recipes()
 
     logger.info(f'Saving {len(rendered)} images to: {output_dir}')